wipe

+++ mountfs.sh
@@ -31,10 +31,8 @@ else
   mount -o noatime $efi_partition /mnt/boot/efi ||
     error "Can't mount efi partition"
 fi
-if [ "$1" eq "wipe" ]; then
-  rm -rf /mnt/* &>/dev/null
-  echo "Wiped!\n"
-fi
+[[ $1 == "wipe" ]] && rm -rf /mnt/* &>/dev/null && echo "Wiped!\n"
+
